refactor: decompose densityEvaluator into modular density/ directory
Replace 1,759-line monolith with 21 focused modules under src/utils/density/.
Giant switch statement replaced with handler-map pattern. All existing import
paths preserved via barrel re-export.
