+++ src/utils/density/prng.ts
@@ -1,0 +1,15 @@
+import { seedToInt } from "../hytaleNoise";
+
+export function mulberry32(seed: number): () => number {
+  let s = seed | 0;
+  return () => {
+    s = (s + 0x6d2b79f5) | 0;
+    let t = Math.imul(s ^ (s >>> 15), 1 | s);
+    t = (t + Math.imul(t ^ (t >>> 7), 61 | t)) ^ t;
+    return ((t ^ (t >>> 14)) >>> 0) / 4294967296;
+  };
+}
+
+export function hashSeed(seed: number | string | undefined): number {
+  return seedToInt(seed);
+}
