+++ src/utils/positionEvaluator.ts
@@ -1,4 +1,5 @@
 import type { Node, Edge } from "@xyflow/react";
+import { mulberry32 } from "./density/prng";
 
 /* ── Types ────────────────────────────────────────────────────────── */
 
@@ -13,18 +14,6 @@ export interface WorldRange {
   maxX: number;
   minZ: number;
   maxZ: number;
-}
-
-/* ── Seeded PRNG ──────────────────────────────────────────────────── */
-
-function mulberry32(seed: number): () => number {
-  let s = seed | 0;
-  return () => {
-    s = (s + 0x6d2b79f5) | 0;
-    let t = Math.imul(s ^ (s >>> 15), 1 | s);
-    t = (t + Math.imul(t ^ (t >>> 7), 61 | t)) ^ t;
-    return ((t ^ (t >>> 14)) >>> 0) / 4294967296;
-  };
 }
 
 /* ── Constants ────────────────────────────────────────────────────── */
